requirements.txt
numpy==2.0.2
pandas==2.2.3
matplotlib==3.9.2
tensorflow==2.15.0
scikit-learn==1.5.2
nibabel==5.3.2
python-dotenv==1.0.1
opencv-python==4.10.0.84
Flask==3.1.0